requirements.txt
google-genai
ollama
chromadb
python-dotenv
sentence-transformers
flask
flask-cors